+++ CHANGELOG.md
@@ -8,6 +8,9 @@
     - !181 improves error message handling, now composing a single error message for _all_ encountered errors and providing axis-specific information on the errors. 
     - !181 extends `set_legend` to allow gathering handles and labels from already existing `matplotlib.legend.Legend` objects.
     - !181 improves docstrings of helper methods to convey more clearly which kinds of arguments are expected.
+- The plotting framework now *experimentally* supports skipping of plots.
+  Skipping is triggered via a custom `SkipPlot` exception, that users may raise in their plot functions.
+  Additionally, the `MultiversePlotCreator` allows skipping a plot if the dimensionality of the associated multiverse is not in a set of expected dimensionalities.
 - Documentation:
     - !187 adds links to the source files from which example code is included into the documentation.
     - !189 improves the names of the introductory guides, as proposed in #190.